Make nd_uint8_t and nd_int8_t arrays, to catch direct references.
This catches direct references, so we can change them to use EXTRACT_U_1
or EXTRACT_S_1.

Also, change some structures to use the nd_ types that weren't already
using them.

Then make the appropriate EXTRACT_{U,S}_1() changes.

+++ print-ppp.c
@@ -961,7 +961,7 @@ handle_pap(netdissect_options *ndo,
 		return;
 	}
 	length = len;
-	if (length < (p - p0)) {
+	if (length < (size_t)(p - p0)) {
 		ND_PRINT((ndo, ", length %u < PAP header length", length));
 		return;
 	}
